+++ pkg/client/client.go
@@ -20,7 +20,7 @@ package client
 type Client interface {
 	Init() error
 	Close() error
-	Call(req *Request) (resp Response, err error)
+	Call(req *Request) (resp *Response, err error)
 
 	// MappingParams mapping param, uri, query, body ...
 	MappingParams(req *Request) (types []string, reqData []interface{}, err error)
